requirements.txt
aiosqlite==0.20.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.3
bleach==6.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.121.2
filelock==3.19.1
fsspec==2025.10.0
greenlet==3.1.1
h11==0.16.0
hf-xet==1.2.0
httptools==0.7.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
limits==4.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
passlib==1.7.4
pillow==10.4.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.7.0
pydantic_core==2.41.5
pyotp==2.9.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.3
qrcode==7.4.2
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.3.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.49.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.5.1
tqdm==4.67.1
transformers==4.57.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
wrapt==2.0.1
